(#33): remove --checkpoint option
Apparently the "parent" option to a checkpoints xml file is readonly
according to the documentation.  Libvirt versions will simply ignore
that option and use the last checkpoint existing as parent,
so its actually not possible to force a parent checkpoint.

Remove this option from virtndbdbackup, it has
never had any effect.