Introduce check for checkpoints in chain not created by virtnbdbackup
As discussed during issue #33, the "parent" option to a checkpoint
created is readonly currently. Thus, virtnbdbackup can not force
an parent checkpoint during creation of the actual checkpoints.

This means we cannot ensure that the complete checkpoint chain
we are based on for the current incremental backup is entirely
created by us, like so:

 > virsh checkpoint-list vm1 --tree
 > virtnbdbackup.0                                                    r
 >   |
 >   +- virtnbdbackup.1
 >       |
 >       +- virtnbdbackup.2
 >           |
 >           +- not-our-checkpoint

The delta for "not our checkpoint" would never be saved because
we dont know about this checkpoint.

Now virtnbdbackup checks for checkpoints which might have been
created by users or third party applications and exits with
error if incremental backup is attempted.