Fix inc backup of offline domain
Offline backup would allways use the qemu executable
to query extents, but qemu-img map will allways return
the extents of the base allocation context. As such,
the incremental backup of an offline vm would be turned
into an full backup.

Change this behavior and exit gracefully if the -q
option is used during inc backup.