(#35) add differencial backup option
It is now possible to create differencial backups too (-l diff).
An differencial backup will not create a new checkpoint but will
save the differences to the last existing checkpoint.

The last checkpoint can either be a full, or incremental backup.
Incremental and differencial backups as such can be mixed, as
such it is up to the user how big the differencial backups
will grow.

One can create an base full backup and then only create further
differencial backups, which will then grow in size until another
incremental or full backup is executed.