add option to specify log directory
This is helpful when invoking `virtnbdrestore` via `subprocess.Popen` or
similar without a full shell environment, which will spawn it with `/`
as the working directory.
Previously it would then attempt to open `/virtnbdrestore.log` which
usually fails as an unprivileged user.

To deal with this, this commit introduces `--logdir` and it defaults to
`os.environ['HOME']` which should ensure proper permissions by
default.