handle lost CBT case (#272)
* handle lost CBT case

It is legal that CBT inside QEMU could be lost f.e. once the power on
the node would be lost or QEMU will be crashed. Dirty bitmaps are
stored inside QCOW2 as auto-clear feature and are removed on every
start when the image is unclean. This is done as CBTs are kept in the
memory and are written only during normal QEMU shutdown process.

This situation should be handled by the backup software. Backup chains
should not be broken and the only normal way to handle this would be
to make full backup and store it inside incremental chain. Letting
resolving this by the end-user would be not friendly to the end-user.

The condition is rare and we should just create full backup in this case.
Handle this respectively in the code.

Signed-off-by: Denis V. Lunev <den@openvz.org>

* check: remove redundant args.offline=True in vmstate() helper

This assignment is redundant as this branch is reachable only iff
args.offline == True, which was checked two lines above.

Signed-off-by: Denis V. Lunev <den@openvz.org>

* basic test for incremental backup with missed CBT

The test is based on the nosparsedetect test.

Signed-off-by: Denis V. Lunev <den@openvz.org>

---------

Signed-off-by: Denis V. Lunev <den@openvz.org>